# openapi.yaml
openapi: 3.0.3
info:
  title: Fudo Challenge API
  version: 1.0.0
  description: Simple Rack-based API with authentication and async product creation

paths:
  /auth:
    post:
      description: Given a valid username and password it authenticates the user and returns a token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: Invalid credentials
        '400':
          description: Invalid json

  /products:
    get:
      description: List products that are in memory
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Returns all products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/Product'
        '401':
          description: Unauthorized

    post:
      description: Create a product asynchronously
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '202':
          description: Product creation scheduled
        '401':
          description: Unauthorized

components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token